Generate summary headlines in an overnight background task
Extract the headline-generation pipeline out of SummarySection into a
shared, headless SummaryHeadlineGenerator used by both the Home UI and a
new background task, so there is a single source of truth.

Add SummaryBackfillScheduler, a charging-gated BGProcessingTask that is
submitted 30 minutes after each background feed refresh finishes. It
pre-generates and caches headlines for each summary kind that is within
its display window and not already cached, so the Home cards load from
cache instead of generating on app start.

Register the task identifier in Info.plist.
