+++ App/App+BackgroundTasks.swift
@@ -50,6 +50,13 @@ extension SakuraRSSApp {
             guard let task = task as? BGProcessingTask else { return }
             self.handleImageBackfill(task: task)
         }
+        BGTaskScheduler.shared.register(
+            forTaskWithIdentifier: SummaryBackfillScheduler.taskIdentifier,
+            using: nil
+        ) { task in
+            guard let task = task as? BGProcessingTask else { return }
+            self.handleSummaryBackfill(task: task)
+        }
     }
 
     /// Submits a `BGAppRefreshTaskRequest` per category so each one gets its
@@ -122,6 +129,7 @@ extension SakuraRSSApp {
                 "BackgroundRefresh",
                 "handleAppRefresh end category=\(category.rawValue) cancelled=\(refreshTask.isCancelled)"
             )
+            SummaryBackfillScheduler.schedule()
             completion.complete(success: !refreshTask.isCancelled)
         }
     }
@@ -291,6 +299,28 @@ extension SakuraRSSApp {
         Task {
             _ = await work.value
             log("NighttimeBackfill", "handleImageBackfill end cancelled=\(work.isCancelled)")
+            completion.complete(success: !work.isCancelled)
+        }
+    }
+
+    nonisolated func handleSummaryBackfill(task: BGProcessingTask) {
+        log("SummaryBackfill", "handleSummaryBackfill begin")
+
+        let completion = BackgroundTaskCompletion(task: task)
+
+        let work = Task {
+            await SummaryBackfillScheduler.runBackfill(isCancelled: { Task.isCancelled })
+        }
+
+        task.expirationHandler = {
+            log("SummaryBackfill", "handleSummaryBackfill expired")
+            work.cancel()
+            completion.complete(success: false)
+        }
+
+        Task {
+            _ = await work.value
+            log("SummaryBackfill", "handleSummaryBackfill end cancelled=\(work.isCancelled)")
             completion.complete(success: !work.isCancelled)
         }
     }
