+++ App/Info.plist
@@ -14,6 +14,7 @@
 		<string>com.tsubuzaki.SakuraRSS.AutomaticCleanup</string>
 		<string>com.tsubuzaki.SakuraRSS.NLPBackfill</string>
 		<string>com.tsubuzaki.SakuraRSS.ImageBackfill</string>
+		<string>com.tsubuzaki.SakuraRSS.SummaryBackfill</string>
 	</array>
 	<key>CFBundleDocumentTypes</key>
 	<array>
